added componentDidUpdate() for Home component
