+++ client/components/search.jsx
@@ -32,7 +32,7 @@ export default class Search extends React.Component {
   }
 
   componentDidUpdate(prevProps) {
-    if (prevProps !== this.props) {
+    if (this.props !== prevProps) {
       const searchKeyWords = this.props.value.replaceAll(' ', '+');
       const url = `https://www.googleapis.com/books/v1/volumes?q=${searchKeyWords}&projection=full&key=${process.env.GOOGLE_BOOKS_API_KEY}`;
       const request = {
